requirements.txt
BAC0>=20.02.20
paho-mqtt>=1.5.0
ruamel.yaml>=0.15
